requirements.txt
aiohttp==3.6.2
pyjwt==1.7.1
pypandoc  # only for converting README.md
